+++ swagger.js
@@ -24,7 +24,7 @@ const doc = {
 };
 
 const outputFile = './docs/docs.json';
-const endpointsFiles = ['./http/app.js'];
+const endpointsFiles = ['./app.js'];
 
 /* NOTE: if you use the express Router, you must pass in the
    'endpointsFiles' only the root file where the route starts,
